🚧 Add logic, compile options, and import to use AvxToNeon for ARM compilation
Note that this relies on the library being available within
`ThirdPartyLibs`, which is not yet managed by the repo or build scripts!

With these changes and the library manually placed in its location,
building fails due to missing `_mm256_loadu_ps` function, which is
documented as supported but appears to be missing from the library!