Improves mass calibration workflow
Introduces a new version of the mass calibration setter,
utilizing lazy loading for improved performance.
This version truncates the scan info list to focus on a representative
subset for calibration, optimizing resource usage and calibration
speed. It also uses a parallel loader for turboXIC to improve
calibration speed.

Disables a test that needs to be rewritten.