Makes turboXicMS1 initialization optional
Allows the CandidateScorertron and TargetDecoyCandidatePairScoretron to function without requiring turboXicMS1 initialization.

This change makes the MS1 feature optional, improving flexibility and allowing the algorithm to run in scenarios where MS1 data is unavailable or not needed.

Adds conditional checks around turboXicMS1 usage to prevent errors when it is not initialized.