Implements Pythia DDA workflow
Adds the initial implementation of the Pythia DDA workflow. This includes the workflow class, initialization logic with parameter validation and frag library reading, and a basic file processing structure. The workflow integrates with MsReaderMzMLLazyLoad for reading MS data files. The older file is deleted. Adds to CMakeLists.txt and adds a basic test.

Also adds debug logs to CandidateScoresFeatureManager.cpp